+++ detox/test/index.ios.js
@@ -14,91 +14,55 @@ import {
   TouchableOpacity
 } from 'react-native';
 import _ from 'lodash';
+import * as Screens from './src/Screens'
 
-const STRESSFUL_STRING_LENGTH = 48000; // Min: 32000
-const STRESSFUL_EVENTS_COUNT = 570; // Min: 380
-
-function buildStringByLength(length) {
-  str = "";
-  charcode = 65;
-  for (let i=0; i < length; i++) {
-    str += String.fromCharCode(charcode);
-    charcode ++;
-    if (charcode == 91) charcode = 65;
-  }
-  return str;
-}
+import SanityScreen from './src/Screens/SanityScreen'
+import StressScreen from './src/Screens/StressScreen'
 
 class example extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      greeting: undefined,
-      passToBridge: undefined
+      screen: undefined
     };
   }
 
-  renderTestButton(label, onPress) {
+  renderScreenButton(screen) {
     return (
-      <TouchableOpacity onPress={onPress}>
-        <Text style={{color: 'blue', marginBottom: 20}}>{label}</Text>
+      <TouchableOpacity onPress={() => {
+        this.setState({
+          screen
+        });
+      }}>
+        <Text style={{color: 'blue', marginBottom: 20}}>{screen}</Text>
       </TouchableOpacity>
     )
   }
 
   render() {
-    if (this.state.greeting) return this.renderAfterButton();
-    return (
-      <View style={{flex: 1, paddingTop: 20, justifyContent: 'center', alignItems: 'center'}}>
-        <Text style={{fontSize: 25, marginBottom: 30}}>
-          Welcome
-        </Text>
-        {this.renderTestButton('Say Hello', this.onButtonPress.bind(this, 'Hello'))}
-        {this.renderTestButton('Say World', this.onButtonPress.bind(this, 'World'))}
-        {this.renderTestButton('Bridge Stress', this.bridgeStressButtonPressed.bind(this, 'Hello World'))}
-        {this.renderTestButton('Events Stress', this.eventsStressButtonPressed.bind(this, 'Hello World'))}
-      </View>
-    );
-  }
-  renderAfterButton() {
-    return (
-      <View style={{flex: 1, paddingTop: 20, justifyContent: 'center', alignItems: 'center'}}>
-        <Text style={{fontSize: 10, width: 0, height: 0}}>
-          Bridge: {this.state.passToBridge}
-        </Text>
-        <Text style={{fontSize: 25}}>
-          {this.state.greeting}!!!
-        </Text>
-      </View>
-    );
-  }
-
-  onButtonPress(greeting) {
-    this.setState({
-      greeting: greeting
-    });
-  }
-
-  bridgeStressButtonPressed(greeting) {
-    this.onButtonPress(greeting)
-
-    const data = buildStringByLength(STRESSFUL_STRING_LENGTH);
-    this.setState({
-      passToBridge: data,
-      greeting: greeting
-    });
-  }
-
-  eventsStressButtonPressed(greeting) {
-    // Stress:
-    for (let i =0; i < STRESSFUL_EVENTS_COUNT; i++) {
-      let myString = "";
-      setImmediate(() => { myString = buildStringByLength(1000) });
+    if (!this.state.screen) {
+      return (
+        <View style={{flex: 1, paddingTop: 20, justifyContent: 'center', alignItems: 'center'}}>
+          <Text style={{fontSize: 25, marginBottom: 30}}>
+            Please Choose a test
+          </Text>
+          {this.renderScreenButton(Screens.SANITY_SCREEN)}
+          {this.renderScreenButton(Screens.STRESS_SCREEN)}
+        </View>
+      );
     }
 
-    setImmediate(() => { this.onButtonPress(greeting) });
+    switch (this.state.screen) {
+      case Screens.SANITY_SCREEN:
+        return (
+          <SanityScreen></SanityScreen>
+        );
+      case Screens.STRESS_SCREEN:
+        return (
+          <StressScreen></StressScreen>
+        );
+    }
   }
-
 }
 
 AppRegistry.registerComponent('example', () => example);
