+++ detox/test/src/Screens/index.js
@@ -1,0 +1,2 @@
+export const SANITY_SCREEN = "SanityScreen";
+export const STRESS_SCREEN = "StressScreen";
